requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization  
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Network Analysis
networkx>=3.1.0

# Data Processing
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Development Tools
jupyter>=1.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Documentation
sphinx>=7.0.0
